Add logging to discord-services.js

+++ discord-services.js
@@ -1,4 +1,5 @@
 const Discord = require('discord.js');
+const winston = require('winston');
 const BotGuild = require('./db/mongo/BotGuild');
 
 // where hackers join the wait list to talk to a sponsor
@@ -20,6 +21,7 @@ module.exports.sponsorCategory = sponsorCategory;
  * @param {Discord.Snowflake} role - role ID to check for
  */
 function checkForRole(member, role) {
+    winston.loggers.get(member.guild.id).verbose(`A role check was requested. Role ID: ${role}. Member ID: ${member.id}`);
     return member.roles.cache.has(role);
 }
 module.exports.checkForRole = checkForRole;
@@ -37,7 +39,7 @@ async function sendMsgToChannel(channel, userId, message, timeout = 0) {
     let msg = await channel.send('<@' + userId + '> ' + message);
 
     if (timeout) msg.delete({timeout: timeout * 1000}); // convert to milliseconds
-
+    winston.loggers.get(channel.guild.id).verbose(`A message has been sent to the channel ${channel.name} for the user with id ${userId} ${timeout === 0 ? 'with no timeout requested' : 'with a ' + timeout + ' second timeout.'}`);
     return msg;
 }
 module.exports.sendMsgToChannel = sendMsgToChannel;
@@ -52,15 +54,20 @@ module.exports.sendMsgToChannel = sendMsgToChannel;
  */
 async function sendMessageToMember(member, message, isDelete = false) {
     return await member.send(message).then(msg => {
+        winston.loggers.get(member.guild.id).verbose(`A DM message was sent to user with id ${member.id}.`);
         if (isDelete === true) {
             msg.delete({timeout: 60000})
         }
         return msg;
     }).catch(async error => {
         if (error.code === 50007) {
+            winston.loggers.get(member.guild.id).warning(`A DM message was sent to user with id ${member.id} but failed, he has been asked to fix this problem!`);
             let botGuild;
             if (member?.guild) botGuild = await BotGuild.findById(member.guild.id);
-            else throw Error(`I could not help ${member.id} due to not finding the guild he is trying to access. I need a member and not a user!`);
+            else {
+                winston.loggers.get(member.guild.id).error(`While trying to help a user to get my DMs I could not find a botGuild for which this member is in. I could not help him!`);
+                throw Error(`I could not help ${member.id} due to not finding the guild he is trying to access. I need a member and not a user!`);
+            }
 
             member.guild.channels.resolve(botGuild.channelIDs.botSupportChannel).send('<@' + member.id + '> I couldn\'t reach you :(. Please turn on server DMs, explained in this link: https://support.discord.com/hc/en-us/articles/217916488-Blocking-Privacy-Settings-');
         } else {
@@ -117,14 +124,16 @@ module.exports.sendEmbedToMember = sendEmbedToMember;
  * @param {Discord.RoleResolvable} addRole - the role to add to the member
  */
 function addRoleToMember(member, addRole) {
+    let role = member.guild.roles.resolve(addRole);
     member.roles.add(addRole).catch(error => {
-        console.log(error);
         // try one more time
         member.roles.add(addRole).catch(error => {
             // now send error to admins
-            discordLog(member.guild, '@everyone The member <@' + member.user.id + '> did not get the role <@&' + member.guild.roles.cache.get(addRole).id +'> please help me!');
+            discordLog(member.guild, '@everyone The member <@' + member.user.id + '> did not get the role <@&' + role.id +'> please help me!');
+            winston.loggers.get(member.guild.id).error(`Could not give the member with id ${member.id} the role ${role.name} with id ${role.id}. The following error ocurred: ${error.name} - ${error.message}.`);
         });
     });
+    winston.loggers.get(member.guild.id).verbose(`A member with id ${member.id} was given the role ${role.name} with id ${role.id}`);
 }
 module.exports.addRoleToMember = addRoleToMember;
 
@@ -134,13 +143,16 @@ module.exports.addRoleToMember = addRoleToMember;
  * @param {Discord.RoleResolvable} removeRole - the role to add to the member
  */
 function removeRolToMember(member, removeRole) {
+    let role = member.guild.roles.resolve(removeRole);
     member.roles.remove(removeRole).catch(error => {
         // try one more time
         member.roles.remove(removeRole).catch(error => {
             // now send error to admins
             discordLog(member.guild, '@everyone The member <@' + member.user.id + '> did not loose the role ' + member.guild.roles.cache.get(removeRole).id + ', please help me!');
+            winston.loggers.get(member.guild.id).error(`Could not remove the member with id ${member.id} the role ${role.name} with id ${role.id}. The following error ocurred: ${error.name} - ${error.message}.`);
         });
     });
+    winston.loggers.get(member.guild.id).verbose(`A member with id ${member.id} lost the role ${role.name} with id ${role.id}`);
 }
 module.exports.removeRolToMember = removeRolToMember;
 
@@ -164,7 +176,11 @@ module.exports.replaceRoleToMember = replaceRoleToMember;
  */
 async function discordLog(guild, message) {
     let botGuild = await BotGuild.findById(guild.id);
-    if (botGuild.channelIDs.adminLog) guild.channels.cache.get(botGuild.channelIDs.adminLog).send(message);
+    if (botGuild?.channelIDs?.adminLog) {
+        guild.channels.cache.get(botGuild.channelIDs.adminLog).send(message);
+        winston.loggers.get(guild.id).silly(`The following was logged to discord: ${message}`);
+    }
+    else winston.loggers.get(guild.id).error(`I was not able to log something to discord!! I could not find the botGuild or the adminLog channel!`);
 }
 module.exports.discordLog = discordLog;
 
@@ -176,6 +192,7 @@ module.exports.discordLog = discordLog;
 async function replyAndDelete(message, reply) {
     var msg = await message.reply(reply);
     msg.delete({timeout: 5000});
+    winston.loggers.get(guild.id).verbose(`A message with id ${message.id} is being replied to and then the reply is being deleted.`);
 }
 module.exports.replyAndDelete = replyAndDelete;
 
@@ -186,20 +203,27 @@ module.exports.replyAndDelete = replyAndDelete;
  */
 function deleteMessage(message, timeout = 0) {
     if (!message.deleted && message.deletable &&  message.channel.type != 'dm') {
+        winston.loggers.get(guild.id).verbose(`A message with id ${message.id} in the guild channel ${message.channel.name} with id ${message.channel.id} was deleted.`);
         message.delete({timeout: timeout});
     } else if (message.channel.type === 'dm' && message.author.bot) {
+        winston.loggers.get(guild.id).verbose(`A message with id ${message.id} in a DM channel with user id ${message.channel.recipient.id} from the bot was deleted.`);
         message.delete({timeout: timeout})
+    } else {
+        winston.loggers.get(guild.id).warning(`A message with id ${message.id} in a DM channel from user with id ${message.author.id} tried to be deleted but was not possible.`);
     }
 }
 module.exports.deleteMessage = deleteMessage;
 
 /**
  * Delete the given channel if it is not deleted already
- * @param {Discord.Channel} channel 
+ * @param {Discord.TextChannel} channel 
  */
 async function deleteChannel(channel) {
-    if (!channel.deleted) {
+    if (!channel.deleted && channel.deletable) {
+        winston.loggers.get(channel.guild.id).verbose(`The channel ${channel.name} with id ${channel.id} was deleted.`);
         await channel.delete().catch(console.error);
+    } else {
+        winston.loggers.get(guild.id).warning(`The channel ${channel?.name} with id ${channel?.id} tried to be deleted but was not possible!`);
     }
 }
 module.exports.deleteChannel = deleteChannel;
@@ -209,6 +233,7 @@ module.exports.deleteChannel = deleteChannel;
  * @returns {String} - hex color
  */
 function randomColor() {
+    winston.loggers.get('main').silly(`A random color has been used!`);
     return Math.floor(Math.random()*16777215).toString(16);
 }
 module.exports.randomColor = randomColor;
@@ -219,6 +244,8 @@ module.exports.randomColor = randomColor;
  * @returns {Boolean} true if valid email, false otherwise
  */
 function validateEmail(email) {
+    winston.loggers.get('main').silly(`An email has been validated!`);
+
     // make email lowercase
     email = email.toLowerCase();
 
@@ -227,6 +254,7 @@ function validateEmail(email) {
 
     // let user know he has used the command incorrectly and exit
     if (email === '' || !re.test(email)) {
+        
         return false;
     } else {
         return true;
